feature: treat empty default as valid (#77)
# Pull Request

## Issue

Issue #, if available: N/A

## Description

Description of changes: 

For Terraform there are some cases where we want to either provide an
empty value or a valid string. E.g. an empty string or a subscription
id. The existing implementation did not handle this, so added a
parameter to support this scenario.

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the projects associated license.