Update README.md
While trying to use the ALZ module I kept getting the error: 

command was found in the module 'ALZ', but the module could not be
loaded. For more information, run 'Import-Module ALZ'.
objectNotFound CouldNotAutoloadMatchingModule

and eve though i had updated to PS 7 - it kept showing as 5. I needed to set the alias so everything would work.