enhancement: add terraform installer (#92)
# Pull Request

## Issue

Not applicable.

## Description

This is an enhancement to ensure the dependency on a recent version of
Terraform is handled. This PR checks for and installs the latest version
into a local path if needed. The install script supports multiple os and
architecture.

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the projects associated license.