Enhancement add terraform rerun and destroy (#93)
# Pull Request

## Issue

Issue #, if available:
https://github.com/Azure/alz-terraform-accelerator/issues/70

## Description

Description of changes:

This PR adds re-run, upgrade and destroy features for Terraform. These
features are designed to simplify usage of the accelerator, provide any
easy way to apply changes, fix issues and upgrade to a newer version.
The destroy will make it much easier to clean up test environments.

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the projects associated license.