bug: fix missing auto approve flag (#97)
# Pull Request

## Issue

Issue #, if available: N/A

## Description

This PR fixes a bug where the auto approve parameter is not being passed
to the function that gets the variables. This is causing the module to
prompt for user input when it should not.

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the projects associated license.