Add config file specific code (#100)
# Pull Request

## Issue

Issue #, if available: N/A

## Description

Description of changes:

This PR adds the ability to pass a config file path that is then
included in the repository. This is to be more concise for users of the
accelerator and only prompt for a config file if they are using a
starter module that needs one. Also enables them to call the config file
whatever name they desire.

This change has been tested and testing evidence can be found in the
sister PR here:
https://github.com/Azure/alz-terraform-accelerator/pull/89

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the projects associated license.