enhancement: relocate bootstrap modules (#105)
# Pull Request

## Issue

Issue #, if available: N/A

## Description

This PR is a significant refactor to de-couple bootstrap and starter
modules. It moves to a configuration based approach for shared inputs
and have many other enhancements to the user experience.

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the projects associated license.