Updates to support telemetry and better prompts (#115)
# Pull Request

## Issue

#114 

## Description

This PR makes no significant functional changes other than improving the
prompts users see for IaC, bootstrao and starter.

The primary update is to pull the bootstrap and starter from named
artefacts. By doing this we can get some usage telemetry from the GitHub
API.

Note that changes have been made to the starter and bootstrap repos for
Terraform to accommodate this, but all those changes are backwards
compatible and already released. They are ready to go with these changes
now.

## Testing

- E2E tests have be run for this branch here:
https://github.com/Azure/accelerator-bootstrap-modules/actions/runs/9414281487
- Manual testing of Terraform and Bicep happy paths has been completed
for prompt testing.


## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the projects associated license.