chore: deprecate old docs site (#272)
# Pull Request

## Description

Depecate the old accelerator docs site and add a link to the home page.

Temporarily retain the complete multi region config files for any in
flight deployments. They will be removed in February.

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the projects associated license.