fix: update ALZ module version checks to use Get-InstalledPSResource (#465)
# Pull Request

## Issue

Issue #, if available:
https://github.com/Azure/ALZ-PowerShell-Module/issues/461

## Description

Description of changes:

This pull request updates the way the ALZ module is checked and
retrieved in the `Test-Tooling` function, modernizing the approach to
use the latest PowerShell resource management cmdlets for better
compatibility and reliability.

**Module management modernization:**

* Replaced `Get-InstalledModule` with `Get-InstalledPSResource` for
checking the installed ALZ module version, and updated the selection of
properties to include `Name` and `Version`.
* Replaced `Find-Module` with `Find-PSResource` for retrieving the
latest available ALZ module version.

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the projects associated license.

Signed-off-by: Jan Egil Ring <janegilring@microsoft.com>