feat: improve role definition clean up and add exclusion filters (#495)
# Pull Request

## Description

This PR updates the role assignment query to use resouce graph due to a
bug with the az cli functions.

It also adds two filter to specify which child management groups and
which role assignments to delete. These are to enable the use of this
function at scale for e2e testing.

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the projects associated license.